VCF Loader multi with different chroms on a separate file
